feat: Add sealed type hierarchy for location search results
Implements SearchLocationResult sealed interface with record implementations
for CountryResult, StateResult, and CityResult.

- Replaced Location class with typed result hierarchy
- Added pattern matching support in search service
- Updated test coverage for sealed type verification
- Removed Location.java class, not needed
- Restructured model packages into lookup and search for clarity
- Rebuilt binary location4j.bin with new package names
- Improved type safety and API clarity

Resolves #41
