+++ buildtools/src/main/java/com/tomaytotomato/JsonToBinaryConverter.java
@@ -3,9 +3,9 @@ package com.tomaytotomato;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
-import com.tomaytotomato.location4j.model.City;
-import com.tomaytotomato.location4j.model.Country;
-import com.tomaytotomato.location4j.model.State;
+import com.tomaytotomato.location4j.model.lookup.City;
+import com.tomaytotomato.location4j.model.lookup.Country;
+import com.tomaytotomato.location4j.model.lookup.State;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
