+++ location4j/src/test/java/com/tomaytotomato/location4j/usecase/FindCountriesByStateTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.tomaytotomato.location4j.usecase.lookup.LocationService;
-import com.tomaytotomato.location4j.model.Country;
+import com.tomaytotomato.location4j.model.lookup.Country;
 import com.tomaytotomato.location4j.usecase.lookup.FindCountry;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
