+++ location4j/src/test/java/com/tomaytotomato/location4j/usecase/FindCountryTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.tomaytotomato.location4j.usecase.lookup.LocationService;
-import com.tomaytotomato.location4j.model.Country;
+import com.tomaytotomato.location4j.model.lookup.Country;
 import com.tomaytotomato.location4j.usecase.lookup.FindCountry;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
