+++ location4j/src/test/java/com/tomaytotomato/location4j/usecase/SearchLocationTest.java
@@ -1,16 +1,21 @@
 package com.tomaytotomato.location4j.usecase;
 
-import com.tomaytotomato.location4j.usecase.search.SearchLocationService;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import com.tomaytotomato.location4j.aliases.DefaultLocationAliases;
 import com.tomaytotomato.location4j.loader.DefaultCountriesDataLoaderImpl;
-import com.tomaytotomato.location4j.mapper.DefaultLocationMapper;
+import com.tomaytotomato.location4j.mapper.DefaultSearchLocationResultMapper;
+import com.tomaytotomato.location4j.model.search.CityResult;
+import com.tomaytotomato.location4j.model.search.CountryResult;
+import com.tomaytotomato.location4j.model.search.SearchLocationResult;
+import com.tomaytotomato.location4j.model.search.StateResult;
 import com.tomaytotomato.location4j.text.normaliser.DefaultTextNormaliser;
 import com.tomaytotomato.location4j.text.tokeniser.DefaultTextTokeniser;
 import com.tomaytotomato.location4j.usecase.search.SearchLocation;
+import com.tomaytotomato.location4j.usecase.search.SearchLocationService;
 import java.util.Arrays;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.Assertions.tuple;
-import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+import java.util.List;
+import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
@@ -25,7 +30,7 @@ class SearchLocationTest {
   public SearchLocationTest() {
     searchLocationService = SearchLocationService.builder()
         .withLocationAliases(new DefaultLocationAliases())
-        .withLocationMapper(new DefaultLocationMapper())
+        .withLocationMapper(new DefaultSearchLocationResultMapper())
         .withCountriesDataLoader(new DefaultCountriesDataLoaderImpl())
         .withTextNormaliser(new DefaultTextNormaliser())
         .withTextTokeniser(new DefaultTextTokeniser())
@@ -35,7 +40,6 @@ class SearchLocationTest {
   @DisplayName("SearchLocation, when null or empty text, then throw exception")
   @Test
   void search_WhenNullOrBlank_ThenThrowException() {
-
     // When Then
     assertThatThrownBy(() -> searchLocationService.search(null)).isInstanceOf(
             IllegalArgumentException.class)
@@ -62,12 +66,15 @@ class SearchLocationTest {
       "United Arab Emirates"
   })
   void search_WhenTextContainsCountryOnly_ThenReturnCountryMatch(String countryName) {
-
     // When
     var result = searchLocationService.search(countryName);
 
     // Then
-    assertThat(result).isNotEmpty().hasSize(1).extracting("countryName").containsOnly(countryName);
+    assertThat(result).isNotEmpty().hasSize(1);
+    assertThat(result.getFirst()).isInstanceOf(CountryResult.class);
+
+    var countryResult = (CountryResult) result.getFirst();
+    assertThat(countryResult.getCountryName()).isEqualTo(countryName);
   }
 
   @DisplayName("SearchLocation, when text contains country ISO2, then return country match")
@@ -87,9 +94,13 @@ class SearchLocationTest {
     var expectedCountries = expectedCountryMatches.split("\\|");
 
     // Then
-    assertThat(result).isNotEmpty().hasSize(expectedCountries.length)
-        .extracting("countryName")
-        .containsAll(Arrays.asList(expectedCountries));
+    assertThat(result).isNotEmpty().hasSize(expectedCountries.length);
+
+    List<String> countryNames = result.stream()
+        .map(SearchLocationResult::getCountryName)
+        .collect(Collectors.toList());
+
+    assertThat(countryNames).containsAll(Arrays.asList(expectedCountries));
   }
 
   @DisplayName("SearchLocation, when text contains country ISO3, then return country match")
@@ -108,9 +119,12 @@ class SearchLocationTest {
     var result = searchLocationService.search(iso3Code);
 
     // Then
-    assertThat(result).isNotEmpty().hasSize(1)
-        .extracting("countryName", "countryIso3Code")
-        .containsOnly(tuple(expectedCountryName, iso3Code));
+    assertThat(result).isNotEmpty().hasSize(1);
+    assertThat(result.getFirst()).isInstanceOf(CountryResult.class);
+
+    CountryResult countryResult = (CountryResult) result.getFirst();
+    assertThat(countryResult.getCountryName()).isEqualTo(expectedCountryName);
+    assertThat(countryResult.getCountryIso3Code()).isEqualTo(iso3Code);
   }
 
   @DisplayName("SearchLocation, when text contains state and country name, then return single match")
@@ -136,30 +150,60 @@ class SearchLocationTest {
     var result = searchLocationService.search(text);
 
     // Then
-    assertThat(result).isNotEmpty().hasSize(1)
-        .extracting("countryName")
-        .containsOnly(countryName);
+    assertThat(result).isNotEmpty().hasSize(1);
+    assertThat(result.getFirst().getCountryName()).isEqualTo(countryName);
   }
 
+  @ParameterizedTest
+  @DisplayName("When searching with {0}, then returns a CityResult with expected data")
+  @CsvSource(delimiter = '|', value = {
+      "San Francisco, CA, USA|United States|US|USA|California|San Francisco",
+      "KY, Glasgow USA|United States|US|USA|Kentucky|Glasgow",
+      "Glasgow Kentucky USA|United States|US|USA|Kentucky|Glasgow",
+      "Glasgow Scotland|United Kingdom|GB|GBR|Scotland|Glasgow"
+  })
+  void search_WithCombinedLocationText_ReturnsCityResult(
+      String searchText,
+      String expectedCountry,
+      String expectedIso2,
+      String expectedIso3,
+      String expectedState,
+      String expectedCity) {
 
-  @Test
-  void search_WhenTextContainsCountryStateandCityName_ThenReturnSingleMatch() {
+    // When
+    var results = searchLocationService.search(searchText);
 
-    // When Then
-    assertThat(searchLocationService.search("San Francisco, CA, USA")).isNotEmpty().hasSize(1)
-        .extracting("countryName", "countryIso2Code", "countryIso3Code", "stateName", "city")
-        .containsExactly(tuple("United States", "US", "USA", "California", "San Francisco"));
+    // Then
+    assertThat(results).isNotEmpty().hasSize(1);
+    assertThat(results.getFirst()).isInstanceOf(CityResult.class);
 
-    assertThat(searchLocationService.search("KY, Glasgow USA")).isNotEmpty().hasSize(1)
-        .extracting("countryName", "countryIso2Code", "countryIso3Code", "stateName", "city")
-        .containsExactly(tuple("United States", "US", "USA", "Kentucky", "Glasgow"));
+    CityResult cityResult = (CityResult) results.getFirst();
+    assertThat(cityResult.getCountryName()).isEqualTo(expectedCountry);
+    assertThat(cityResult.getCountryIso2Code()).isEqualTo(expectedIso2);
+    assertThat(cityResult.getCountryIso3Code()).isEqualTo(expectedIso3);
+    assertThat(cityResult.getStateName()).isEqualTo(expectedState);
+    assertThat(cityResult.getCityName()).isEqualTo(expectedCity);
+  }
 
-    assertThat(searchLocationService.search("Glasgow Kentucky USA")).isNotEmpty().hasSize(1)
-        .extracting("countryName", "countryIso2Code", "countryIso3Code", "stateName", "city")
-        .containsExactly(tuple("United States", "US", "USA", "Kentucky", "Glasgow"));
+  @ParameterizedTest
+  @DisplayName("Pattern matching with sealed types produces correct descriptions")
+  @CsvSource(delimiter = '|', value = {
+      "San Francisco, CA, USA|City: San Francisco, California, United States",
+      "CA, United States|State: California, United States",
+      "United States|Country: United States"
+  })
+  void patternMatchingWithDifferentLocationTypes(String searchText, String expectedDescription) {
+    var result = searchLocationService.search(searchText).getFirst();
 
-    assertThat(searchLocationService.search("Glasgow Scotland")).isNotEmpty().hasSize(1)
-        .extracting("countryName", "countryIso2Code", "countryIso3Code", "stateName", "city")
-        .containsExactly(tuple("United Kingdom", "GB", "GBR", "Scotland", "Glasgow"));
+    var description = switch (result) {
+      case CountryResult countryResult -> "Country: " + countryResult.getCountryName();
+      case StateResult stateResult ->
+          "State: " + stateResult.getStateName() + ", " + stateResult.getCountryName();
+      case CityResult cityResult -> "City: " + cityResult.getCityName() + ", " +
+          cityResult.getStateName() + ", " +
+          cityResult.getCountryName();
+    };
+
+    assertThat(description).isEqualTo(expectedDescription);
   }
 }
